requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
imagehash>=4.2.0